feat: outlet replaced and csv generation action implemented
