+++ src/components/SelectedBooksFlyout.tsx
@@ -1,11 +1,15 @@
+'use client';
+
+import { useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
+import { generateCSVDownload } from '@/app/actions/generate-csv';
 import Button from './Button';
 import { selectSelectedBooks, clearSelectedBooks } from '../store/booksSlice';
-import generateCSVDownloadUrl from '../utils/saveCSV';
 
 function SelectedBooksFlyout() {
   const dispatch = useDispatch();
   const selectedBooks = useSelector(selectSelectedBooks);
+  const [isLoading, setIsLoading] = useState(false);
 
   if (selectedBooks.length === 0) {
     return null;
@@ -15,7 +19,24 @@ function SelectedBooksFlyout() {
     dispatch(clearSelectedBooks());
   };
 
-  const { url, filename } = generateCSVDownloadUrl(selectedBooks);
+  const handleDownload = async () => {
+    setIsLoading(true);
+    try {
+      const { csvContent, filename } = await generateCSVDownload(selectedBooks);
+
+      const blob = new Blob([csvContent], { type: 'text/csv;charset=utf-8;' });
+      const url = URL.createObjectURL(blob);
+      const a = document.createElement('a');
+      a.href = url;
+      a.download = filename;
+      a.click();
+      URL.revokeObjectURL(url);
+    } catch (error) {
+      console.error('Failed to generate CSV:', error);
+    } finally {
+      setIsLoading(false);
+    }
+  };
 
   return (
     <div className="fixed right-4 bottom-4 rounded-lg border border-gray-200 bg-white p-4 shadow-lg dark:border-gray-700 dark:bg-gray-800 dark:shadow-gray-700/50">
@@ -26,9 +47,9 @@ function SelectedBooksFlyout() {
         <Button color="secondary" onClick={handleClearSelection}>
           Unselect all
         </Button>
-        <a href={url} download={filename}>
-          <Button>Download</Button>
-        </a>
+        <Button onClick={() => void handleDownload()} disabled={isLoading}>
+          {isLoading ? 'Generating...' : 'Download'}
+        </Button>
       </div>
     </div>
   );
